requirements.txt
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
Flask==3.0.3
Flask-Caching==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pyodbc==5.1.0
python-dotenv==1.0.1
schedule==1.2.2
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
